src/novalis/requirements.txt
appdirs==1.4.4
asgiref==3.3.4
cuser==0.0.4
distlib==0.3.1
Django==3.1.7
django-crispy-forms==1.11.2
django-filebrowser==3.13.2
django-grappelli==2.14.3
django-js-asset==1.2.2
django-tinymce==3.3.0
django-widget-tweaks==1.4.8
filelock==3.0.12
middleware==1.2.3
Pillow==8.2.0
pyserial==3.5
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
virtualenv==20.4.3
